requirements.txt
opencv-python>=4.8.0
openai>=1.0.0
Pillow>=9.0.0
numpy==1.24.3
requests>=2.25.0
pyinstaller>=6.0.0
